+++ 03-admin-todos/src/todos/actions/todo-actions.ts
@@ -3,12 +3,16 @@
 import { revalidatePath } from "next/cache";
 import { Todo } from "@prisma/client";
 import prisma from "@/lib/prisma";
+import { getUserSessionServer } from "@/auth/actions/auth-actions";
 
 export const toggleTodo = async (
   id: string,
   complete: boolean
 ): Promise<Todo> => {
-  const todo = await prisma.todo.findFirst({ where: { id } });
+  const user = await getUserSessionServer();
+  if (!user) throw new Error("User not found");
+
+  const todo = await prisma.todo.findFirst({ where: { id, userId: user.id } });
   if (!todo) throw `TODO con id ${id} no encontrado`;
 
   const updatedTodo = await prisma.todo.update({
@@ -23,7 +27,12 @@ export const toggleTodo = async (
 // : Promise<Todo>
 export const addTodo = async (description: string) => {
   try {
-    const todo = await prisma.todo.create({ data: { description } });
+    const user = await getUserSessionServer();
+    if (!user) throw new Error("User not found");
+
+    const todo = await prisma.todo.create({
+      data: { description, userId: user.id },
+    });
 
     revalidatePath("/dsahboard/server-todos");
     return todo;
@@ -35,8 +44,11 @@ export const addTodo = async (description: string) => {
 
 export const deleteTodos = async (): Promise<void> => {
   try {
+    const user = await getUserSessionServer();
+    if (!user) throw new Error("User not found");
+
     await prisma.todo.deleteMany({
-      where: { complete: true },
+      where: { complete: true, userId: user.id },
     });
 
     revalidatePath("/dsahboard/server-todos");
